proper fix for ast_group_t change
git-svn-id: https://origsvn.digium.com/svn/asterisk/branches/1.4@43978 65c4cc65-6c06-0410-ace0-fbb531ad65f3

+++ main/cli.c
@@ -891,8 +891,8 @@ static int handle_showchan(int fd, int argc, char *argv[])
 		"        Context: %s\n"
 		"      Extension: %s\n"
 		"       Priority: %d\n"
-		"     Call Group: %d\n"
-		"   Pickup Group: %d\n"
+		"     Call Group: %llu\n"
+		"   Pickup Group: %llu\n"
 		"    Application: %s\n"
 		"           Data: %s\n"
 		"    Blocking in: %s\n",
@@ -910,7 +910,7 @@ static int handle_showchan(int fd, int argc, char *argv[])
 		c->fout & ~DEBUGCHAN_FLAG, (c->fout & DEBUGCHAN_FLAG) ? " (DEBUGGED)" : "",
 		(long)c->whentohangup,
 		cdrtime, c->_bridge ? c->_bridge->name : "<none>", ast_bridged_channel(c) ? ast_bridged_channel(c)->name : "<none>", 
-		c->context, c->exten, c->priority, (int) c->callgroup, (int) c->pickupgroup, ( c->appl ? c->appl : "(N/A)" ),
+		c->context, c->exten, c->priority, c->callgroup, c->pickupgroup, ( c->appl ? c->appl : "(N/A)" ),
 		( c-> data ? S_OR(c->data, "(Empty)") : "(None)"),
 		(ast_test_flag(c, AST_FLAG_BLOCKING) ? c->blockproc : "(Not Blocking)"));
